pyproject.toml
[project]
name = "exam-result-summarizer-notifier"
version = "0.1.0"
description = "Flask-based Exam Result Summarizer with LLM feedback and email notifications"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "flask>=2.3.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "huggingface-hub>=0.16.0",
    "torch>=2.0.0",
    "python-dotenv>=1.0.0",
    "flask-session>=0.5.0",
    "werkzeug>=2.3.0",
]
